requirements.txt
fastapi
uvicorn
prometheus_client
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation
opentelemetry-exporter-otlp
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-requests
opentelemetry-semantic-conventions
opentelemetry-instrumentation-logging